@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.4 <3.0.0'
+  sdk: '>=3.4.3 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -38,16 +38,22 @@ dependencies:
   intl: ^0.17.0
   nb_utils: ^4.6.24
   url_launcher: ^6.1.6
-  google_fonts: ^3.0.1
+  google_fonts: ^6.2.1
+
+  #API CLIENT
+  retrofit: '>=4.0.0 <5.0.0'
+  logger: any  #for logging purpose
+  json_annotation: ^4.8.1
 
   # STATE MANAGEMENT
   get: ^4.6.5
+  animations: ^2.0.11
 
   # IMAGE
   cached_network_image: ^3.2.2
 
   # UI
-  flutter_slidable: 0.6.0
+  flutter_slidable: ^3.1.0
   flutter_svg: ^1.1.5
   carousel_slider: ^4.1.1
   flutter_layout_grid: ^2.0.0
@@ -56,6 +62,8 @@ dependencies:
 
   # Common
   html: ^0.15.0
+  equatable: ^2.0.5
+  pretty_dio_logger: ^1.3.1
 
 dev_dependencies:
   flutter_test:
@@ -67,6 +75,9 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^2.0.0
+  retrofit_generator: '>=7.0.0 <8.0.0'
+  build_runner: '>=2.3.0 <4.0.0'
+  json_serializable: ^6.6.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
