@@ -1,264 +1,175 @@
 
+import 'package:json_annotation/json_annotation.dart';
+
 import 'model_category.dart';
 
 class ModelProduct {
   int? id;
   String? name;
+  // ignore: non_constant_identifier_names
+  String? dateCreated;
 
   // ignore: non_constant_identifier_names
-  String? date_created;
 
   // ignore: non_constant_identifier_names
-  String? date_created_gmt;
 
   // ignore: non_constant_identifier_names
-  String? date_modified;
-
-  // ignore: non_constant_identifier_names
-  String? date_modified_gmt;
+  // String? date_modified_gmt;
   String? type;
   String? status;
-  bool? featured;
+  // bool? featured;
 
   // ignore: non_constant_identifier_names
-  String? catalog_visibility;
+  // String? catalog_visibility;
   String? description;
 
   // ignore: non_constant_identifier_names
-  String? short_description;
-  String? sku;
+  // String? short_description;
+  // String? sku;
   String? price;
 
   // ignore: non_constant_identifier_names
   String? regular_price;
 
   // ignore: non_constant_identifier_names
-  String? sale_price;
+  String? salePrice;
 
   // ignore: non_constant_identifier_names
-  String? price_html;
+  // String? price_html;
 
   // ignore: non_constant_identifier_names
-  bool? on_sale;
-  bool? purchasable;
+  // bool? on_sale;
+  // bool? purchasable;
 
   // ignore: non_constant_identifier_names
-  int? total_sales;
-  bool? virtual;
-  bool? downloadable;
+  // int? total_sales;
+  // bool? virtual;
+  // bool? downloadable;
 
 
   // ignore: non_constant_identifier_names
-  String? external_url;
+  // String? external_url;
 
   // ignore: non_constant_identifier_names
-  String? button_text;
+  // String? button_text;
 
   // ignore: non_constant_identifier_names
-  String? tax_status;
+  // String? tax_status;
 
   // ignore: non_constant_identifier_names
-  String? tax_class;
+  // String? tax_class;
 
   // ignore: non_constant_identifier_names
-  bool? manage_stock;
+  // bool? manage_stock;
 
   // ignore: non_constant_identifier_names
-  int? stock_quantity;
+  int? stockQuantity;
 
   // ignore: non_constant_identifier_names
-  String? stock_status;
-  String? backorders;
+  String? stockStatus;
+  // String? backorders;
 
   // ignore: non_constant_identifier_names
-  bool? backorders_allowed;
-  bool? backordered;
+  // bool? backorders_allowed;
+  // bool? backordered;
 
   // ignore: non_constant_identifier_names
-  bool? sold_individually;
-  String? weight;
-  Dimensions? dimensions;
+  // bool? sold_individually;
+  // String? weight;
+  // Dimensions? dimensions;
 
-  // ignore: non_constant_identifier_names
-  bool? shipping_required;
+  // // ignore: non_constant_identifier_names
+  // bool? shipping_required;
+  //
+  // // ignore: non_constant_identifier_names
+  // bool? shipping_taxable;
+  //
+  // // ignore: non_constant_identifier_names
+  // String? shipping_class;
+  //
+  // // ignore: non_constant_identifier_names
+  // int? shipping_class_id;
 
-  // ignore: non_constant_identifier_names
-  bool? shipping_taxable;
-
-  // ignore: non_constant_identifier_names
-  String? shipping_class;
-
-  // ignore: non_constant_identifier_names
-  int? shipping_class_id;
-
-  // ignore: non_constant_identifier_names
-  bool? reviews_allowed;
-
-  // ignore: non_constant_identifier_names
-  String? average_rating;
-
-  // ignore: non_constant_identifier_names
-  int? rating_count;
-
-  // ignore: non_constant_identifier_names
-  List<dynamic>? related_ids;
-
-  // ignore: non_constant_identifier_names
-  List<dynamic>? upsell_ids;
-
-  // ignore: non_constant_identifier_names
-  List<dynamic>? cross_sell_ids;
-
-  // ignore: non_constant_identifier_names
-  int? parent_id;
-
-  // ignore: non_constant_identifier_names
-  String? purchase_note;
+  // // ignore: non_constant_identifier_names
+  // bool? reviews_allowed;
+  //
+  // // ignore: non_constant_identifier_names
+  // String? average_rating;
+  //
+  // // ignore: non_constant_identifier_names
+  // int? rating_count;
+  //
+  // // ignore: non_constant_identifier_names
+  // List<dynamic>? related_ids;
+  //
+  // // ignore: non_constant_identifier_names
+  // List<dynamic>? upsell_ids;
+  //
+  // // ignore: non_constant_identifier_names
+  // List<dynamic>? cross_sell_ids;
+  //
+  // // ignore: non_constant_identifier_names
+  // int? parent_id;
+  //
+  // // ignore: non_constant_identifier_names
+  // String? purchase_note;
   List<ModelCategory>? categories;
-  List<dynamic>? tags;
+  // List<dynamic>? tags;
   List<ShImage>? images;
-  List<Attribute>? attributes;
-
-  // ignore: non_constant_identifier_names
-  List<dynamic>? default_attributes;
+  // List<Attribute>? attributes;
+  //
+  // // ignore: non_constant_identifier_names
+  // List<dynamic>? default_attributes;
 
   ModelProduct(
       {this.id,
       this.name,
       // ignore: non_constant_identifier_names
-      this.date_created,
+      this.dateCreated,
       // ignore: non_constant_identifier_names
-      this.date_created_gmt,
       // ignore: non_constant_identifier_names
-      this.date_modified,
       // ignore: non_constant_identifier_names
-      this.date_modified_gmt,
       this.type,
       this.status,
-      this.featured,
       // ignore: non_constant_identifier_names
-      this.catalog_visibility,
+
       this.description,
       // ignore: non_constant_identifier_names
-      this.short_description,
-      this.sku,
       this.price,
       // ignore: non_constant_identifier_names
       this.regular_price,
       // ignore: non_constant_identifier_names
-      this.sale_price,
+      this.salePrice,
       // ignore: non_constant_identifier_names
-      this.price_html,
       // ignore: non_constant_identifier_names
-      this.on_sale,
-      this.purchasable,
       // ignore: non_constant_identifier_names
-      this.total_sales,
-      this.virtual,
-      this.downloadable,
-
+      // this.external_url,
       // ignore: non_constant_identifier_names
-      this.external_url,
+      this.stockQuantity,
       // ignore: non_constant_identifier_names
-      this.button_text,
-      // ignore: non_constant_identifier_names
-      this.tax_status,
-      // ignore: non_constant_identifier_names
-      this.tax_class,
-      // ignore: non_constant_identifier_names
-      this.manage_stock,
-      // ignore: non_constant_identifier_names
-      this.stock_quantity,
-      // ignore: non_constant_identifier_names
-      this.stock_status,
-      this.backorders,
-      // ignore: non_constant_identifier_names
-      this.backorders_allowed,
-      this.backordered,
-      // ignore: non_constant_identifier_names
-      this.sold_individually,
-      this.weight,
-      this.dimensions,
-      // ignore: non_constant_identifier_names
-      this.shipping_required,
-      // ignore: non_constant_identifier_names
-      this.shipping_taxable,
-      // ignore: non_constant_identifier_names
-      this.shipping_class,
-      // ignore: non_constant_identifier_names
-      this.shipping_class_id,
-      // ignore: non_constant_identifier_names
-      this.reviews_allowed,
-      // ignore: non_constant_identifier_names
-      this.average_rating,
-      // ignore: non_constant_identifier_names
-      this.rating_count,
-      // ignore: non_constant_identifier_names
-      this.related_ids,
-      // ignore: non_constant_identifier_names
-      this.upsell_ids,
-      // ignore: non_constant_identifier_names
-      this.cross_sell_ids,
-      // ignore: non_constant_identifier_names
-      this.parent_id,
-      // ignore: non_constant_identifier_names
-      this.purchase_note,
+      this.stockStatus,
       this.categories,
-      this.tags,
+      // this.tags,
       this.images,
-      this.attributes,
-      // ignore: non_constant_identifier_names
-      this.default_attributes});
+      });
 
   factory ModelProduct.fromJson(Map<String, dynamic> json) {
     return ModelProduct(
       id: json['id'],
       name: json['name'],
-      date_created: json['date_created'],
-      date_created_gmt: json['date_created_gmt'],
-      date_modified: json['date_modified'],
-      date_modified_gmt: json['date_modified_gmt'],
+      dateCreated: json['date_created'],
       type: json['type'],
       status: json['status'],
-      featured: json['featured'],
-      catalog_visibility: json['catalog_visibility'],
       description: json['description'],
-      short_description: json['short_description'],
-      sku: json['sku'],
       price: json['price'],
       regular_price: json['regular_price'],
-      sale_price: json['sale_price'],
-      price_html: json['price_html'],
-      on_sale: json['on_sale'],
-      purchasable: json['purchasable'],
-      total_sales: json['total_sales'],
-      virtual: json['virtual'],
-      downloadable: json['downloadable'],
-      external_url: json['external_url'],
-      button_text: json['button_text'],
-      tax_status: json['tax_status'],
-      tax_class: json['tax_class'],
-      manage_stock: json['manage_stock'],
-      stock_quantity: json['stock_quantity'],
-      stock_status: json['stock_status'],
-      backorders: json['backorders'],
-      backorders_allowed: json['backorders_allowed'],
-      backordered: json['backordered'],
-      sold_individually: json['sold_individually'],
-      weight: json['weight'],
-      dimensions: json['dimensions'] != null ? Dimensions.fromJson(json['dimensions']) : null,
-      shipping_required: json['shipping_required'],
-      shipping_taxable: json['shipping_taxable'],
-      shipping_class: json['shipping_class'],
-      shipping_class_id: json['shipping_class_id'],
-      reviews_allowed: json['reviews_allowed'],
-      average_rating: json['average_rating'],
-      rating_count: json['rating_count'],
-      parent_id: json['parent_id'],
-      purchase_note: json['purchase_note'],
+      salePrice: json['sale_price'],
+      // external_url: json['external_url'],
+      stockQuantity: json['stock_quantity'],
+      stockStatus: json['stock_status'],
       categories: json['categories'] != null ? (json['categories'] as List).map((i) => ModelCategory.fromJson(i)).toList() : null,
       images: json['images'] != null ? (json['images'] as List).map((i) => ShImage.fromJson(i)).toList() : null,
-      attributes: json['attributes'] != null ? (json['attributes'] as List).map((i) => Attribute.fromJson(i)).toList() : null,
+      // attributes: json['attributes'] != null ? (json['attributes'] as List).map((i) => Attribute.fromJson(i)).toList() : null,
     );
   }
 
@@ -266,116 +177,42 @@ class ModelProduct {
     final Map<String, dynamic> data = <String, dynamic>{};
     data['id'] = id;
     data['name'] = name;
-    data['date_created'] = date_created;
-    data['date_created_gmt'] = date_created_gmt;
-    data['date_modified'] = date_modified;
-    data['date_modified_gmt'] = date_modified_gmt;
+    data['date_created'] = dateCreated;
     data['type'] = type;
     data['status'] = status;
-    data['featured'] = featured;
-    data['catalog_visibility'] = catalog_visibility;
     data['description'] = description;
-    data['short_description'] = short_description;
-    data['sku'] = sku;
     data['price'] = price;
     data['regular_price'] = regular_price;
-    data['sale_price'] = sale_price;
-    data['price_html'] = price_html;
-    data['on_sale'] = on_sale;
-    data['purchasable'] = purchasable;
-    data['total_sales'] = total_sales;
-    data['virtual'] = virtual;
-    data['downloadable'] = downloadable;
-    data['external_url'] = external_url;
-    data['button_text'] = button_text;
-    data['tax_status'] = tax_status;
-    data['tax_class'] = tax_class;
-    data['manage_stock'] = manage_stock;
-    data['stock_quantity'] = stock_quantity;
-    data['stock_status'] = stock_status;
-    data['backorders'] = backorders;
-    data['backorders_allowed'] = backorders_allowed;
-    data['backordered'] = backordered;
-    data['sold_individually'] = sold_individually;
-    data['weight'] = weight;
-    if (dimensions != null) {
-      data['dimensions'] = dimensions!.toJson();
-    }
-    data['shipping_required'] = shipping_required;
-    data['shipping_taxable'] = shipping_taxable;
-    data['shipping_class'] = shipping_class;
-    data['shipping_class_id'] = shipping_class_id;
-    data['reviews_allowed'] = reviews_allowed;
-    data['average_rating'] = average_rating;
-    data['rating_count'] = rating_count;
-
-    data['parent_id'] = parent_id;
-    data['purchase_note'] = purchase_note;
+    data['sale_price'] = salePrice;
+    // data['external_url'] = external_url;
+    data['stock_quantity'] = stockQuantity;
+    data['stock_status'] = stockStatus;
     if (categories != null) {
       data['categories'] = categories!.map((v) => v.toJson()).toList();
     }
-    if (tags != null) {
-      data['tags'] = tags!.map((v) => v.toJson()).toList();
-    }
     if (images != null) {
       data['images'] = images!.map((v) => v.toJson()).toList();
-    }
-    if (attributes != null) {
-      data['attributes'] = attributes!.map((v) => v.toJson()).toList();
-    }
-    if (default_attributes != null) {
-      data['default_attributes'] =
-          default_attributes!.map((v) => v.toJson()).toList();
     }
     return data;
   }
 }
 
-
 class ShImage {
   int? id;
-
-  // ignore: non_constant_identifier_names
-  String? date_created;
-
-  // ignore: non_constant_identifier_names
-  String? date_created_gmt;
-
-  // ignore: non_constant_identifier_names
-  String? date_modified;
-
-  // ignore: non_constant_identifier_names
-  String? date_modified_gmt;
   String? src;
-  String? name;
-  String? alt;
-
-  // ignore: non_constant_identifier_names
-  ShImage({this.id, this.date_created, this.date_created_gmt, this.date_modified, this.date_modified_gmt, this.src, this.name, this.alt});
+  ShImage({this.id, this.src});
 
   factory ShImage.fromJson(Map<String, dynamic> json) {
     return ShImage(
       id: json['id'],
-      date_created: json['date_created'],
-      date_created_gmt: json['date_created_gmt'],
-      date_modified: json['date_modified'],
-      date_modified_gmt: json['date_modified_gmt'],
       src: json['src'],
-      name: json['name'],
-      alt: json['alt'],
     );
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = <String, dynamic>{};
     data['id'] = id;
-    data['date_created'] = date_created;
-    data['date_created_gmt'] = date_created_gmt;
-    data['date_modified'] = date_modified;
-    data['date_modified_gmt'] = date_modified_gmt;
     data['src'] = src;
-    data['name'] = name;
-    data['alt'] = alt;
     return data;
   }
 }
