@@ -1,3 +1,5 @@
+import 'package:ecommerce_responsive/api/product_api_impl.dart';
+import 'package:ecommerce_responsive/models/product_response.dart';
 import 'package:ecommerce_responsive/screens/screen_search_product.dart';
 import 'package:ecommerce_responsive/screens/side_drawer.dart';
 import 'package:ecommerce_responsive/utils/widgets/dots_indicator/src/dots_decorator.dart';
@@ -17,6 +19,7 @@ import 'package:ecommerce_responsive/utils/strings_constant.dart';
 import 'package:ecommerce_responsive/utils/common_widget.dart';
 import 'package:ecommerce_responsive/utils/widgets/app_widget.dart';
 
+import '../utils/common_widget_new.dart';
 import '../utils/widgets/appbar.dart';
 
 class ScreenHome extends StatefulWidget {
@@ -31,8 +34,9 @@ class ScreenHome extends StatefulWidget {
 class ScreenHomeState extends State<ScreenHome> {
   List<ModelCategory> categoryList = [];
   List<String> banners = [];
-  List<ModelProduct> newestProducts = [];
-  List<ModelProduct> featuredProducts = [];
+  List<ProductResponse> newestProducts = [];
+  List<ProductResponse> products = [];
+  List<ProductResponse> featuredProducts = [];
   var position = 0;
   var colors = [sh_cat_1, sh_cat_2, sh_cat_3, sh_cat_4, sh_cat_5];
 
@@ -51,13 +55,15 @@ class ScreenHomeState extends State<ScreenHome> {
     }).catchError((error) {
       toasty(context, error);
     });
-    List<ModelProduct> products = await loadProducts();
-    List<ModelProduct> featured = [];
-    products.forEach((product) {
+    // List<ModelProduct> products = await loadProducts();
+    List<ProductResponse>? products = await ApiIpml.productApi.getAllProducts() ?? [];
+    List<ProductResponse> featured = [];
+    products.map((product) {
       if (product.featured!) {
         featured.add(product);
       }
     });
+    /// banner
     List<String> banner = [];
     for (var i = 1; i < 7; i++) {
       banner.add("assets/images/shopx/img/products/banners/b$i.jpg");
@@ -74,108 +80,122 @@ class ScreenHomeState extends State<ScreenHome> {
 
   @override
   Widget build(BuildContext context) {
-
     return Scaffold(
-      appBar: const DefaultAppBar(title: "ShopX",showLeading: true,),
-      drawer: context.isPhone()? const SideDrawer():null,
+      appBar: const DefaultAppBar(
+        title: "ShopX",
+        showLeading: true,
+      ),
+      drawer: context.isPhone() ? const SideDrawer() : null,
       body: newestProducts.isNotEmpty
           ? SafeArea(
-            child: SingleChildScrollView(
+              child: SingleChildScrollView(
                 child: Column(
                   children: <Widget>[
                     Stack(
-                      alignment: const Alignment(0.9,-0.9),
+                      alignment: const Alignment(0.9, -0.9),
                       children: [
                         SizedBox(
                           child: ConstrainedBox(
-                            constraints: const BoxConstraints(maxHeight: 280, minWidth: 500),
-                            child: StatefulBuilder(
-                                builder: (context, setState) {
-                                return Stack(
-                                  alignment: Alignment.bottomCenter,
-                                  children: <Widget>[
-                                    PageView.builder(
-                                      itemCount: banners.length,
-                                      itemBuilder: (context, index) {
-                                        return Image.asset(banners[index], fit: BoxFit.contain);
-                                      },
-                                      onPageChanged: (index) {
-                                        setState(() {
-                                          position = index;
-                                        });
-                                      },
-                                    ),
-                                    Padding(
-                                      padding: const EdgeInsets.all(8.0),
-                                      child: DotsIndicator(
-                                        dotsCount: banners.length,
-                                        position: position,
-                                        decorator: const DotsDecorator(
-                                          color: sh_view_color,
-                                          activeColor: sh_colorPrimary,
-                                          size: Size.square(7.0),
-                                          activeSize: Size.square(10.0),
-                                          spacing: EdgeInsets.all(3),
-                                        ),
+                            constraints: const BoxConstraints(
+                                maxHeight: 280, minWidth: 500),
+                            child:
+                                StatefulBuilder(builder: (context, setState) {
+                              return Stack(
+                                alignment: Alignment.bottomCenter,
+                                children: <Widget>[
+                                  PageView.builder(
+                                    itemCount: banners.length,
+                                    itemBuilder: (context, index) {
+                                      return Image.asset(banners[index],
+                                          fit: BoxFit.contain);
+                                    },
+                                    onPageChanged: (index) {
+                                      setState(() {
+                                        position = index;
+                                      });
+                                    },
+                                  ),
+                                  Padding(
+                                    padding: const EdgeInsets.all(8.0),
+                                    child: DotsIndicator(
+                                      dotsCount: banners.length,
+                                      position: position,
+                                      decorator: const DotsDecorator(
+                                        color: sh_view_color,
+                                        activeColor: sh_colorPrimary,
+                                        size: Size.square(7.0),
+                                        activeSize: Size.square(10.0),
+                                        spacing: EdgeInsets.all(3),
                                       ),
                                     ),
-                                  ],
-                                );
-                              }
-                            ),
+                                  ),
+                                ],
+                              );
+                            }),
                           ),
                         ),
-
-                        Card(
-                            elevation: 4,
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(15.0),
-                            ),child: const Padding(
-                          padding: EdgeInsets.all(8.0),
-                          child: Icon(Icons.search),
-                        )
-                        ).onTap(()=> Get.toNamed(ScreenSearchProduct.tag))
                       ],
                     ),
 
                     ///Category list
-                    if(!context.isPhone())
-                    Center(
-                      child: ListView.builder(
-                        itemCount: categoryList.length,
-                        shrinkWrap: true,
-                        scrollDirection: Axis.horizontal,
-                        padding: const EdgeInsets.only(left: spacing_standard, right: spacing_standard,top: spacing_standard_new),
-                        itemBuilder: (BuildContext context, int index) {
-                          return Container(
-                            margin: const EdgeInsets.only(left: spacing_standard, right: spacing_standard),
-                            child: Column(
-                              children: <Widget>[
-                                Container(
-                                  padding: const EdgeInsets.all(spacing_middle),
-                                  decoration: BoxDecoration(shape: BoxShape.circle, color: colors[index % colors.length]),
-                                  child: Image.asset(categoryList[index].image!, width: 20, color: sh_white),
-                                ),
-                                const SizedBox(height: spacing_control),
-                                text(categoryList[index].name, textColor: colors[index % colors.length], fontFamily: fontMedium,fontSize: 22)
-                              ],
-                            ),
-                          ).onTap(() {
-                            ScreenCategory(category: categoryList[index]).launch(context);
-                          });
-                        },
-                      ),
-                    ).withHeight(100),
+                    if (!context.isPhone())
+                      Center(
+                        child: ListView.builder(
+                          itemCount: categoryList.length,
+                          shrinkWrap: true,
+                          scrollDirection: Axis.horizontal,
+                          padding: const EdgeInsets.only(
+                              left: spacing_standard,
+                              right: spacing_standard,
+                              top: spacing_standard_new),
+                          itemBuilder: (BuildContext context, int index) {
+                            return Container(
+                              margin: const EdgeInsets.only(
+                                  left: spacing_standard,
+                                  right: spacing_standard),
+                              child: Column(
+                                children: <Widget>[
+                                  Container(
+                                    padding:
+                                        const EdgeInsets.all(spacing_middle),
+                                    decoration: BoxDecoration(
+                                        shape: BoxShape.circle,
+                                        color: colors[index % colors.length]),
+                                    child: Image.asset(
+                                        categoryList[index].image!,
+                                        width: 20,
+                                        color: sh_white),
+                                  ),
+                                  const SizedBox(height: spacing_control),
+                                  text(categoryList[index].name,
+                                      textColor: colors[index % colors.length],
+                                      fontFamily: fontMedium,
+                                      fontSize: 22)
+                                ],
+                              ),
+                            ).onTap(() {
+                              ScreenCategory(category: categoryList[index])
+                                  .launch(context);
+                            });
+                          },
+                        ),
+                      ).withHeight(100),
 
                     ///Newest Product
-                    horizontalHeading(sh_lbl_newest_product, callback: () {
-                      ScreenViewAllProduct(products: newestProducts, title: sh_lbl_newest_product).launch(context);
+                    horizontalHeadingV2(sh_lbl_newest_product, callback: () {
+                      ScreenViewAllProduct(
+                              products: newestProducts,
+                              title: sh_lbl_newest_product)
+                          .launch(context);
                     }),
-                    ProductHorizontalList(newestProducts),
+                    ProductHorizontalListV2(newestProducts),
 
                     const SizedBox(height: spacing_standard_new),
-                    horizontalHeading(sh_lbl_Featured, callback: () {
-                      ScreenViewAllProduct(products: featuredProducts, title: sh_lbl_Featured).launch(context);
+                    horizontalHeadingV2(sh_lbl_Featured, callback: () {
+                      ScreenViewAllProduct(
+                              products: featuredProducts,
+                              title: sh_lbl_Featured)
+                          .launch(context);
                     }),
                     ProductHorizontalList(featuredProducts),
                     const SizedBox(height: 60),
@@ -184,7 +204,7 @@ class ScreenHomeState extends State<ScreenHome> {
                   ],
                 ),
               ),
-          )
+            )
           : Container(),
     );
   }
